+++ src/main/java/umc/cockple/demo/domain/exercise/service/ExerciseQueryService.java
@@ -7,25 +7,15 @@ import org.springframework.transaction.annotation.Transactional;
 import umc.cockple.demo.domain.exercise.converter.ExerciseConverter;
 import umc.cockple.demo.domain.exercise.domain.Exercise;
 import umc.cockple.demo.domain.exercise.domain.ExerciseAddr;
-import umc.cockple.demo.domain.exercise.domain.Guest;
 import umc.cockple.demo.domain.exercise.dto.*;
-import umc.cockple.demo.domain.exercise.dto.ExerciseDetailDTO.ParticipantInfo;
-import umc.cockple.demo.domain.exercise.exception.ExerciseErrorCode;
-import umc.cockple.demo.domain.exercise.exception.ExerciseException;
+import umc.cockple.demo.domain.exercise.service.support.ExerciseParticipantInfoAssembler;
 import umc.cockple.demo.domain.exercise.service.support.ExerciseParticipantReader;
 import umc.cockple.demo.domain.exercise.service.support.ExerciseReader;
-import umc.cockple.demo.domain.exercise.service.support.GuestReader;
 import umc.cockple.demo.domain.member.service.support.MemberLookupService;
 import umc.cockple.demo.domain.member.domain.Member;
-import umc.cockple.demo.domain.member.domain.MemberExercise;
 import umc.cockple.demo.domain.party.domain.Party;
-import umc.cockple.demo.global.enums.Gender;
-import umc.cockple.demo.global.enums.Level;
-import umc.cockple.demo.global.enums.Role;
 
-import java.time.LocalDate;
 import java.util.*;
-import java.util.stream.Collectors;
 
 @Service
 @Transactional(readOnly = true)
@@ -34,8 +24,8 @@ import java.util.stream.Collectors;
 public class ExerciseQueryService {
 
     private final ExerciseReader exerciseReader;
-    private final GuestReader guestReader;
     private final ExerciseParticipantReader exerciseParticipantReader;
+    private final ExerciseParticipantInfoAssembler participantInfoAssembler;
     private final MemberLookupService memberLookupService;
 
     private final ExerciseConverter exerciseConverter;
@@ -52,7 +42,7 @@ public class ExerciseQueryService {
 
         ExerciseDetailDTO.ExerciseInfo exerciseInfo = createExerciseInfo(exercise);
 
-        List<ExerciseDetailDTO.ParticipantInfo> allParticipants = getAllSortedParticipants(exerciseId, party);
+        List<ExerciseDetailDTO.ParticipantInfo> allParticipants = participantInfoAssembler.getAllSortedParticipants(exerciseId, party);
         ParticipantGroups groups = splitParticipants(allParticipants, exercise.getMaxCapacity());
 
         ExerciseDetailDTO.ParticipantGroup participantGroup = createParticipantGroup(groups.participants(), exercise.getMaxCapacity());
@@ -61,41 +51,12 @@ public class ExerciseQueryService {
         return exerciseConverter.toDetailResponse(isManager, exerciseInfo, participantGroup, waitingGroup);
     }
 
-    public ExerciseMyGuestListDTO.Response getMyInvitedGuests(Long exerciseId, Long memberId) {
-
-        log.info("내가 초대한 게스트 조회 시작 - exerciseId = {}, memberId = {}", exerciseId, memberId);
-
-        Exercise exercise = exerciseReader.findExerciseWithBasicInfoOrThrow(exerciseId);
-        Member member = memberLookupService.findByIdOrThrow(memberId);
-
-        List<Guest> myGuests = guestReader.findByExerciseIdAndInviterId(exerciseId, memberId);
-
-        if (myGuests.isEmpty()) {
-            log.info("초대한 게스트가 없어 빈 응답 반환 - exerciseId: {}, memberId: {}", exerciseId, memberId);
-            return exerciseConverter.toEmptyGuestListResponse();
-        }
-
-        List<ExerciseDetailDTO.ParticipantInfo> allParticipants = getAllSortedParticipants(exerciseId, exercise.getParty());
-        Map<Long, ExerciseMyGuestListDTO.GuestGroups> guestNumberMap = createGuestNumberMap(allParticipants, exercise.getMaxCapacity());
-
-        String inviterName = member.getMemberName();
-        List<ExerciseMyGuestListDTO.GuestInfo> guestInfoList = buildGuestInfoList(myGuests, guestNumberMap, inviterName);
-
-        ExerciseMyGuestListDTO.GuestStatistics statistics = calculateGuestStatistics(guestInfoList);
-
-        log.info("내가 초대한 게스트 조회 완료 - exerciseId: {}", exerciseId);
-
-        return exerciseConverter.toMyGuestListResponse(statistics, guestInfoList);
-    }
-
     public ExerciseEditDetailDTO.Response getExerciseForEdit(Long exerciseId, Long memberId) {
         log.info("운동 수정용 상세조회 시작 - exerciseId: {}, memberId: {}", exerciseId, memberId);
         Exercise exercise = exerciseReader.findExerciseWithBasicInfoOrThrow(exerciseId);
         log.info("운동 수정용 상세조회 완료 - exerciseId: {}", exerciseId);
         return exerciseConverter.toEditDetailResponse(exercise);
     }
-
-    // ========== 비즈니스 메서드 ==========
 
     // ========== 비즈니스 메서드 ==========
 
@@ -111,22 +72,6 @@ public class ExerciseQueryService {
                 .buildingName(addr.getBuildingName())
                 .location(addr.getStreetAddr())
                 .build();
-    }
-
-    private List<ExerciseDetailDTO.ParticipantInfo> getAllSortedParticipants(Long exerciseId, Party party) {
-        List<MemberExercise> memberExercises = exerciseParticipantReader.findMemberExercisesWithMemberAndProfile(exerciseId);
-        List<ExerciseDetailDTO.ParticipantInfo> memberParticipants = buildMemberParticipantInfos(memberExercises, party);
-
-        List<Guest> guests = guestReader.findByExerciseId(exerciseId);
-        List<ExerciseDetailDTO.ParticipantInfo> guestParticipants = buildGuestParticipantInfos(guests);
-
-        List<ExerciseDetailDTO.ParticipantInfo> allParticipants = new ArrayList<>();
-        allParticipants.addAll(memberParticipants);
-        allParticipants.addAll(guestParticipants);
-
-        allParticipants.sort(Comparator.comparing(ExerciseDetailDTO.ParticipantInfo::joinedAt));
-
-        return allParticipants;
     }
 
     private ParticipantGroups splitParticipants(
@@ -161,91 +106,6 @@ public class ExerciseQueryService {
                 .womenCount(countByGender(waiting, "FEMALE"))
                 .list(waiting)
                 .build();
-    }
-
-    private Map<Long, ExerciseMyGuestListDTO.GuestGroups> createGuestNumberMap(List<ParticipantInfo> allParticipants, Integer maxCapacity) {
-        Map<Long, ExerciseMyGuestListDTO.GuestGroups> guestNumberMap = new HashMap<>();
-
-        for (int i = 0; i < allParticipants.size(); i++) {
-            ExerciseDetailDTO.ParticipantInfo participant = allParticipants.get(i);
-
-            if ("GUEST".equals(participant.participantType())) {
-                if (i < maxCapacity) {
-                    guestNumberMap.put(participant.participantId(),
-                            ExerciseMyGuestListDTO.GuestGroups.participant(i + 1));
-                } else {
-                    int waitingNumber = i - maxCapacity + 1;
-                    guestNumberMap.put(participant.participantId(),
-                            ExerciseMyGuestListDTO.GuestGroups.waiting(waitingNumber));
-                }
-            }
-        }
-
-        return guestNumberMap;
-    }
-
-    private List<ExerciseMyGuestListDTO.GuestInfo> buildGuestInfoList(
-            List<Guest> myGuests,
-            Map<Long, ExerciseMyGuestListDTO.GuestGroups> guestNumberMap,
-            String inviterName) {
-
-        return myGuests.stream()
-                .map(guest -> exerciseConverter.toGuestInfo(guest, guestNumberMap, inviterName))
-                .toList();
-    }
-
-    private ExerciseMyGuestListDTO.GuestStatistics calculateGuestStatistics(List<ExerciseMyGuestListDTO.GuestInfo> guestInfoList) {
-        int totalCount = guestInfoList.size();
-        int maleCount = (int) guestInfoList.stream()
-                .filter(guest -> guest.gender() == Gender.MALE)
-                .count();
-        int femaleCount = totalCount - maleCount;
-
-        return new ExerciseMyGuestListDTO.GuestStatistics(totalCount, maleCount, femaleCount);
-    }
-
-    // ========== 세부 비즈니스 메서드 ==========
-
-    private List<ParticipantInfo> buildMemberParticipantInfos(List<MemberExercise> memberExercises, Party party) {
-        if (memberExercises.isEmpty()) {
-            return List.of();
-        }
-
-        List<Long> memberIds = memberExercises.stream()
-                .map(me -> me.getMember().getId())
-                .toList();
-
-        Map<Long, Role> partyMemberRoles = exerciseParticipantReader
-                .findMemberRolesByPartyAndMembers(party.getId(), memberIds);
-
-        return memberExercises.stream()
-                .map(me -> {
-                    if (partyMemberRoles.containsKey(me.getMember().getId())) {
-                        return exerciseConverter.toParticipantInfoFromMember(me, partyMemberRoles);
-                    } else {
-                        return exerciseConverter.toParticipantInfoFromExternalMember(me);
-                    }
-                })
-                .toList();
-    }
-
-    private List<ParticipantInfo> buildGuestParticipantInfos(List<Guest> guests) {
-        if (guests.isEmpty()) {
-            return List.of();
-        }
-
-        Set<Long> inviterIds = guests.stream()
-                .map(Guest::getInviterId)
-                .collect(Collectors.toSet());
-
-        Map<Long, String> inviterNames = memberLookupService.findNamesByIds(inviterIds);
-
-        return guests.stream()
-                .map(guest -> {
-                    String inviterName = inviterNames.getOrDefault(guest.getInviterId(), "알 수 없음");
-                    return exerciseConverter.toParticipantInfoFromGuest(guest, inviterName);
-                })
-                .toList();
     }
 
     private List<ExerciseDetailDTO.ParticipantInfo> createParticipantList(
